+++ shared/domain/last-air-episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/lastairepisodes/api/LastAirEpisodeRepository.kt
@@ -4,5 +4,5 @@ import com.thomaskioko.tvmaniac.datasource.cache.AirEpisodesByShowId
 import kotlinx.coroutines.flow.Flow
 
 interface LastAirEpisodeRepository {
-    fun observeAirEpisodes(tvShowId: Int): Flow<List<AirEpisodesByShowId>>
+    fun observeAirEpisodes(tvShowId: Long): Flow<List<AirEpisodesByShowId>>
 }
