+++ shared/domain/show/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/discover/api/interactor/UpdateFollowingInteractor.kt
@@ -19,6 +19,6 @@ class UpdateFollowingInteractor constructor(
 }
 
 data class UpdateShowParams(
-    val showId: Int,
+    val showId: Long,
     val addToWatchList: Boolean
 )
